@@ -24,7 +24,7 @@ import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { ProductCapComponent } from './product-cap/product-cap.component';
 import { ProductCapAmountComponent } from './product-cap-amount/product-cap-amount.component';
 import { SharedModule } from '../../shared/shared.module';
-import { MatProgressSpinner } from '@angular/material/progress-spinner';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { CoreModule } from '../../core/core.module';
 
 
@@ -53,11 +53,11 @@ import { CoreModule } from '../../core/core.module';
     MatSelectModule,
     MatDialogModule,
     MatButtonModule,
-    QuillModule.forRoot(),
+    QuillModule,
     TranslateModule,
     MatSlideToggleModule,
     SharedModule,
-    MatProgressSpinner,
+    MatProgressSpinnerModule,
     CoreModule
   ]
 })
