@@ -50,13 +50,8 @@ namespace MEligibilityPlatform.Infrastructure.Repository
         public List<UserInfo> GetUserByRoleId(int roleId, ApiResponse<List<UserGetModel>> users)
         {
             // Creates a LINQ query to join user roles with users and entities
-            var userRoles = _context.UserRoles
-             .Where(x => x.RoleId == roleId)
-             .ToList();
-                    var query =
-                from ur in userRoles
-                join u in users.Data
-                    on ur.UserId equals u.Id
+            var userRoles = _context.UserRoles.Where(x => x.RoleId == roleId).ToList();
+            var query = from ur in userRoles join u in users.Data on ur.UserId equals u.Id
         select new UserInfo
         {
             UserId = u.Id,
